requirements.txt
# Core Django
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-filter==25.1

# Database
psycopg2-binary==2.9.10

# Authentication & Permissions
django-guardian==2.4.0
djoser==2.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
python3-openid==3.2.0

# API & Serialization
drf-spectacular==0.28.0
drf-nested-routers==0.94.1
Markdown==3.8

# Content Management
django-summernote==0.8.20.0
django-smart-selects==1.7.2

# Utilities
django-autoslug==1.9.9
django-cors-headers==4.7.0
pillow==11.1.0

# Production Server
gunicorn==23.0.0

# Security & Dependencies
cryptography==44.0.3
oauthlib==3.2.2
requests==2.32.3
requests-oauthlib==2.0.0
PyJWT==2.9.0

# JSON & Schema
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
referencing==0.36.2

# Other Dependencies
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
defusedxml==0.7.1
idna==3.10
inflection==0.5.1
packaging==25.0
pycparser==2.22
PyYAML==6.0.2
rpds-py==0.25.1
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
